Make a new graph

This is a generic function for creating graphs.

Usage

make_(...)

Arguments

...

Parameters, see details below.

Details

make_() is a generic function for creating graphs. For every graph constructor in igraph that has a make_ prefix, there is a corresponding function without the prefix: e.g. for make_ring() there is also ring(), etc.

The same is true for the random graph samplers, i.e. for each constructor with a sample_ prefix, there is a corresponding function without that prefix.

These shorter forms can be used together with make_(). The advantage of this form is that the user can specify constructor modifiers which work with all constructors. E.g. the with_vertex_() modifier adds vertex attributes to the newly created graphs.

See the examples and the various constructor modifiers below.

Related documentation in the C library

simplify, vcount, edges, get_eids, ecount

See also

Examples

r <- make_(ring(10))
#> Error in make_(ring(10)): could not find function "make_"
l <- make_(lattice(c(3, 3, 3)))
#> Error in make_(lattice(c(3, 3, 3))): could not find function "make_"

r2 <- make_(ring(10), with_vertex_(color = "red", name = LETTERS[1:10]))
#> Error in make_(ring(10), with_vertex_(color = "red", name = LETTERS[1:10])): could not find function "make_"
l2 <- make_(lattice(c(3, 3, 3)), with_edge_(weight = 2))
#> Error in make_(lattice(c(3, 3, 3)), with_edge_(weight = 2)): could not find function "make_"

ran <- sample_(degseq(c(3, 3, 3, 3, 3, 3), method = "configuration"), simplified())
#> Error in sample_(degseq(c(3, 3, 3, 3, 3, 3), method = "configuration"),     simplified()): could not find function "sample_"
degree(ran)
#> Error in degree(ran): could not find function "degree"
is_simple(ran)
#> Error in is_simple(ran): could not find function "is_simple"